bugfix: handleFileDelete does not throw exception if derivate for file does not exist fixed warning in handleFileCreated()
git-svn-id: https://server.mycore.de/svn/mycore/trunk@26117 7089d063-bc14-496d-ac90-53de1872eb6c

+++ mycore-base/src/main/java/org/mycore/datamodel/metadata/MCRFileMetaEventHandler.java
@@ -95,6 +95,10 @@ public class MCRFileMetaEventHandler extends MCREventHandlerBase {
     @Override
     protected void handleFileDeleted(MCREvent evt, MCRFile file) {
         MCRObjectID derivateID = MCRObjectID.getInstance(file.getOwnerID());
+        if (!MCRMetadataManager.exists(derivateID)) {
+            LOGGER.warn("Derivate " + derivateID + " from file '" + file + "' does not exist.");
+            return;
+        }
         MCRDerivate derivate = MCRMetadataManager.retrieveMCRDerivate(derivateID);
         MCRObjectDerivate objectDerivate = derivate.getDerivate();
         if (objectDerivate.deleteFileMetaData(file.getAbsolutePath())) {
@@ -119,7 +123,9 @@ public class MCRFileMetaEventHandler extends MCREventHandlerBase {
         }
 
         String derivateId = file.getOwnerID();
-        String urn = MCRURNManager.getURNForFile(derivateId, file.getName());
+        String absolutePath = file.getAbsolutePath();
+        String path = absolutePath.substring(0, absolutePath.lastIndexOf("/") + 1);
+        String urn = MCRURNManager.getURNForFile(derivateId, path, file.getName());
 
         if (urn == null) {
             return;
